FAI-346: changes following rebase
